Add files via upload
It was observed from the analysis of the two west african states(Nigeria and Ghana) that:
1.Year 2021 recorded the highest percentage of observation value for both countries with Ghana being the highest of 616.677684 and Nigeria having 596.414253
2. The Proportion of population using improved drinking water sources is at the highest level in Ghana having 993.268461 while Nigeria has 851.640481
3. It was further observed that joint WHO/UNICEF Monitoring Programme team gave the data source in 2023 with Ghana being the highest.

In General, WHO/UNICEF Joint Monitoring Programme in 2023 gave an estimate of observation value of 1.315191e+06 with in a time period of 2016 with approximate of 136624.436287 observation value. The Proportion of population using improved drinking water sources was observed to be 272486.433345, while Proportion of population using surface water was 7414.215129. Meanwhile, San Marino has the highest number of observation value of 7600.000000, while Croatia has the least of observation.